removed wallet creator class and moved method to Wallet.cs
